# Replace row-wise DefaultPagePartitioner with ColumnarPagePartitioner