--- a/core/trino-main/src/main/java/io/trino/operator/OperatorFactories.java
+++ b/core/trino-main/src/main/java/io/trino/operator/OperatorFactories.java
@@ -13,11 +13,8 @@
  */
 package io.trino.operator;
 
-import io.airlift.units.DataSize;
-import io.trino.execution.buffer.OutputBuffer;
 import io.trino.operator.join.JoinBridgeManager;
 import io.trino.operator.join.LookupSourceFactory;
-import io.trino.spi.predicate.NullableValue;
 import io.trino.spi.type.Type;
 import io.trino.spiller.PartitioningSpillerFactory;
 import io.trino.sql.planner.plan.PlanNodeId;
@@ -84,14 +81,4 @@ public interface OperatorFactories
             OptionalInt totalOperatorsCount,
             PartitioningSpillerFactory partitioningSpillerFactory,
             BlockTypeOperators blockTypeOperators);
-
-    OutputFactory partitionedOutput(
-            TaskContext taskContext,
-            PartitionFunction partitionFunction,
-            List<Integer> partitionChannels,
-            List<Optional<NullableValue>> partitionConstants,
-            boolean replicateNullsAndAny,
-            OptionalInt nullChannel,
-            OutputBuffer outputBuffer,
-            DataSize maxPagePartitioningBufferSize);
 }