--- a/core/trino-main/src/main/java/io/trino/operator/TrinoOperatorFactories.java
+++ b/core/trino-main/src/main/java/io/trino/operator/TrinoOperatorFactories.java
@@ -13,14 +13,11 @@
  */
 package io.trino.operator;
 
-import io.airlift.units.DataSize;
-import io.trino.execution.buffer.OutputBuffer;
 import io.trino.operator.join.JoinBridgeManager;
 import io.trino.operator.join.JoinProbe.JoinProbeFactory;
 import io.trino.operator.join.LookupJoinOperatorFactory;
 import io.trino.operator.join.LookupJoinOperatorFactory.JoinType;
 import io.trino.operator.join.LookupSourceFactory;
-import io.trino.spi.predicate.NullableValue;
 import io.trino.spi.type.Type;
 import io.trino.spiller.PartitioningSpillerFactory;
 import io.trino.sql.planner.plan.PlanNodeId;
@@ -36,7 +33,6 @@ import static io.trino.operator.join.LookupJoinOperatorFactory.JoinType.FULL_OUT
 import static io.trino.operator.join.LookupJoinOperatorFactory.JoinType.INNER;
 import static io.trino.operator.join.LookupJoinOperatorFactory.JoinType.LOOKUP_OUTER;
 import static io.trino.operator.join.LookupJoinOperatorFactory.JoinType.PROBE_OUTER;
-import static io.trino.operator.output.PartitionedOutputOperator.PartitionedOutputFactory;
 
 public class TrinoOperatorFactories
         implements OperatorFactories
@@ -163,27 +159,6 @@ public class TrinoOperatorFactories
                 totalOperatorsCount,
                 partitioningSpillerFactory,
                 blockTypeOperators);
-    }
-
-    @Override
-    public OutputFactory partitionedOutput(
-            TaskContext taskContext,
-            PartitionFunction partitionFunction,
-            List<Integer> partitionChannels,
-            List<Optional<NullableValue>> partitionConstants,
-            boolean replicateNullsAndAny,
-            OptionalInt nullChannel,
-            OutputBuffer outputBuffer,
-            DataSize maxPagePartitioningBufferSize)
-    {
-        return new PartitionedOutputFactory(
-                partitionFunction,
-                partitionChannels,
-                partitionConstants,
-                replicateNullsAndAny,
-                nullChannel,
-                outputBuffer,
-                maxPagePartitioningBufferSize);
     }
 
     private static List<Integer> rangeList(int endExclusive)