… core/trino-main/src/main/java/io/trino/operator/output/PagePartitionerFactory.java
… core/trino-main/src/main/java/io/trino/operator/output/PositionsAppender.java
… core/trino-main/src/main/java/io/trino/operator/output/PositionsAppenderFactory.java
… core/trino-main/src/main/java/io/trino/sql/planner/LocalExecutionPlanner.java
… core/trino-main/src/test/java/io/trino/block/BlockAssertions.java
… core/trino-main/src/test/java/io/trino/operator/output/BenchmarkPartitionedOutputOperator.java
… core/trino-main/src/test/java/io/trino/operator/output/TestPartitionedOutputOperator.java