<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->

## Description

<!-- Elaborate beyond the title of the PR as necessary to help the reviewers and maintainers.-->
Add PagePartitioner implementation that copies pages column-wise.
The implementation switches to the old row-wise approach for cases where it's not beneficial to go column-wise.

#### Benchmarks
#### # Tpch/tpch 
sf1000 orc partitioned
[cpoo_sf1000_orc_part.pdf](https://github.com/trinodb/trino/files/8116670/cpoo_sf1000_orc_part.pdf)

##### JMH
```
before
Benchmark                                   (channelCount)  (enableCompression)  (nullRate)  (partitionCount)  (positionCount)                                           (type)  Mode  Cnt     Score    Error  Units
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                           BIGINT  avgt   10  1573.631 ± 59.258  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                               BIGINT_SKEWED_HASH  avgt   10  1680.272 ± 45.548  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                DICTIONARY_BIGINT  avgt   10  1070.769 ± 26.690  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                       RLE_BIGINT  avgt   10  1175.133 ± 54.468  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192              BIGINT_PARTITION_CHANNEL_20_PERCENT  avgt   10  1033.867 ± 36.286  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_20_PERCENT  avgt   10  1193.587 ± 28.215  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_50_PERCENT  avgt   10  1205.715 ± 50.353  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_80_PERCENT  avgt   10  1184.721 ± 33.965  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192  BIGINT_DICTIONARY_PARTITION_CHANNEL_100_PERCENT  avgt   10  1212.772 ± 37.896  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                             RLE_PARTITION_BIGINT  avgt   10  2039.410 ± 32.965  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                        RLE_PARTITION_NULL_BIGINT  avgt   10   105.032 ±  1.573  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                     LONG_DECIMAL  avgt   10  2035.498 ± 15.467  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          INTEGER  avgt   10  1700.506 ± 38.784  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                         SMALLINT  avgt   10  1590.867 ± 24.805  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          BOOLEAN  avgt   10  1514.346 ± 49.831  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          VARCHAR  avgt   10  3525.402 ± 27.326  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                     ARRAY_BIGINT  avgt   10   773.194 ± 24.335  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                    ARRAY_VARCHAR  avgt   10  1543.097 ±  8.732  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                               ARRAY_ARRAY_BIGINT  avgt   10  2102.331 ± 62.703  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                MAP_BIGINT_BIGINT  avgt   10  1756.148 ±  7.994  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                     MAP_BIGINT_MAP_BIGINT_BIGINT  avgt   10  4770.774 ± 30.147  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                ROW_BIGINT_BIGINT  avgt   10   814.704 ± 17.876  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                    ROW_ARRAY_BIGINT_ARRAY_BIGINT  avgt   10  1685.758 ± 49.136  ms/op

after
Benchmark                                   (channelCount)  (enableCompression)  (nullRate)  (partitionCount)  (positionCount)                                           (type)  Mode  Cnt     Score     Error  Units
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                           BIGINT  avgt   10   991.093 ±  13.055  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                               BIGINT_SKEWED_HASH  avgt   10  1210.839 ±  70.398  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                DICTIONARY_BIGINT  avgt   10   660.113 ±  15.403  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                       RLE_BIGINT  avgt   10   597.849 ±  27.456  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192              BIGINT_PARTITION_CHANNEL_20_PERCENT  avgt   10   665.862 ±  17.479  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_20_PERCENT  avgt   10   613.446 ±  23.912  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_50_PERCENT  avgt   10   614.956 ±  14.279  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192   BIGINT_DICTIONARY_PARTITION_CHANNEL_80_PERCENT  avgt   10   629.094 ±  17.930  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192  BIGINT_DICTIONARY_PARTITION_CHANNEL_100_PERCENT  avgt   10   749.012 ±  52.654  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                             RLE_PARTITION_BIGINT  avgt   10   923.164 ±  11.992  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                        RLE_PARTITION_NULL_BIGINT  avgt   10    49.544 ±   0.654  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                     LONG_DECIMAL  avgt   10  1202.360 ±  19.934  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          INTEGER  avgt   10   966.761 ±  16.209  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                         SMALLINT  avgt   10   839.102 ±   5.342  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          BOOLEAN  avgt   10   885.486 ±  16.853  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                          VARCHAR  avgt   10  2679.467 ±  29.166  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                     ARRAY_BIGINT  avgt   10   732.964 ±  30.083  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                    ARRAY_VARCHAR  avgt   10  1472.893 ± 135.731  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                               ARRAY_ARRAY_BIGINT  avgt   10  1598.612 ±  34.598  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                MAP_BIGINT_BIGINT  avgt   10  1492.575 ±  41.539  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                     MAP_BIGINT_MAP_BIGINT_BIGINT  avgt   10  4384.284 ±  82.801  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                                ROW_BIGINT_BIGINT  avgt   10   736.369 ±  31.977  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192                    ROW_ARRAY_BIGINT_ARRAY_BIGINT  avgt   10  1313.624 ±  13.079  ms/op
```

<!-- Answer the following questions to help reviewers and maintainers
understand this PR's scope at a glance:
-->

> Is this change a fix, improvement, new feature, refactoring, or other?

> Is this a change to the core query engine, a connector, client library, or the SPI interfaces? (be specific)

> How would you describe this change to a non-technical end user or system administrator?

## Related issues, pull requests, and links

<!-- List any issues fixed by this PR, and provide links to other related PRs, upstream release notes, and other useful resources. For example:
* Fixes #issuenumber
* Related documentation in #issuenumber
* [Some release notes](http://usefulinfo.example.com)
-->

<!-- The following sections are filled in by the maintainer with input from the contributor:
Use :white_check_mark: or (x) to signal selection.
-->

## Documentation

( ) No documentation is needed.
( ) Sufficient documentation is included in this PR.
( ) Documentation PR is available with #prnumber.
( ) Documentation issue #issuenumber is filed, and can be handled later.

## Release notes

( ) No release notes entries required.
( ) Release notes entries required with the following suggested text:

```markdown
# Section
* Fix some things. ({issue}`issuenumber`)
```