@electrum 2022-02-24
We don't need the `PagePartitioner` interface since there is only one implementation, so please also remove it by reverting 6058b291762278d10438e3eecc6f3b13734bf93b