@sopel39 2022-03-03
mind automation. Either it's (new) flakynes or something is broken (`TestHiveDynamicPartitionPruningTest`)