@lukasz-stec 2022-03-06
I also ran columnar page partitioner benchmarks on x86 and the gains are even better

|   | label            | TPCH wall time | TPC-DS wall time | TPCH CPU time | TPC-DS CPU time | TPCH wall time % change | TPC-DS wall time % change | TPCH CPU time % change | TPC-DS CPU time % change |
| - | ---------------- | -------------- | ---------------- | ------------- | --------------- | ----------------------- | ------------------------- | ---------------------- | ------------------------ |
| 0 | baseline x86     | 993            | 1254             | 119332        | 154405          | 0                       | 0                         | 0                      | 0                        |
| 1 | columnar poo x86 | 944            | 1208             | 110465        | 146709          | \-4.95                  | \-3.6                     | \-7.43                 | \-4.98                   |

[cpoo_sf1000_orc_part_x86.pdf](https://github.com/trinodb/trino/files/8192610/cpoo_sf1000_orc_part_x86.pdf)